add new class
